requirements.txt
streamlit>=1.37.0
# Snowflake SQLAlchemy 1.6+ supports SQLAlchemy 2.x
snowflake-sqlalchemy>=1.7.0
pandas>=2.2.0
sqlalchemy>=2.0.30
plotly>=5.24.0
python-dotenv>=1.0.0
snowflake-snowpark-python>=1.35.0
pydeck>=0.8.1
numpy>=1.24.0
